Fix DiskIO preview failing after scope idles
The scope seems to go into an 'idle' state after about 20 seconds of inactivity. In this state global vars seem to be freed from memory. This caused an issue if then entering DiskIO's preview as the `data` var is then undefined.

So moved all preview preparation code to the card bindings.